Notification control is not notification card and simple notification is now Notification.
