+++ src/Avalonia.Controls/Notifications/NotificationType.cs
@@ -1,0 +1,10 @@
+﻿namespace Avalonia.Controls.Notifications
+{
+    public enum NotificationType
+    {
+        Information,
+        Success,
+        Warning,
+        Error
+    }
+}
