+++ src/Avalonia.Themes.Default/WindowNotificationManager.xaml
@@ -21,104 +21,6 @@
         </Setter>
     </Style>
 
-    <Style Selector="Notification">
-        <Setter Property="UseLayoutRounding" Value="True"/>
-        <Setter Property="Width" Value="350"/>
-        <Setter Property="FontSize" Value="14"/>
-        <Setter Property="Foreground" Value="White"/>
-        <Setter Property="RenderTransformOrigin" Value="50%,75%"/>
-        <Setter Property="Template">
-            <ControlTemplate>
-                <LayoutTransformControl Name="PART_LayoutTransformControl" UseRenderTransform="True">
-                    <Border Background="{TemplateBinding Background}" 
-                            BorderBrush="{TemplateBinding BorderBrush}" 
-                            BorderThickness="{TemplateBinding BorderThickness}"
-                            Margin="8,8,0,0">
-                        <ContentControl MinHeight="150" Content="{TemplateBinding Content}" />
-                    </Border>
-                </LayoutTransformControl>
-            </ControlTemplate>
-        </Setter>
-
-        <Style.Animations>
-            <Animation Duration="0:0:0.45" Easing="QuadraticEaseIn" FillMode="Forward">
-                <KeyFrame Cue="0%">
-                    <Setter Property="Opacity" Value="0"/>
-                    <Setter Property="TranslateTransform.Y" Value="20"/>
-                    <Setter Property="ScaleTransform.ScaleX" Value="0.85"/>
-                    <Setter Property="ScaleTransform.ScaleY" Value="0.85"/>
-                </KeyFrame>
-                <KeyFrame Cue="30%">
-                    <Setter Property="TranslateTransform.Y" Value="-20"/>
-                </KeyFrame>
-                <KeyFrame Cue="100%">
-                    <Setter Property="Opacity" Value="1"/>
-                    <Setter Property="TranslateTransform.Y" Value="0"/>
-                    <Setter Property="ScaleTransform.ScaleX" Value="1"/>
-                    <Setter Property="ScaleTransform.ScaleY" Value="1"/>
-                </KeyFrame>
-            </Animation>
-        </Style.Animations>
-    </Style>
-
-    <Style Selector="Notification[IsClosing=true] /template/ LayoutTransformControl#PART_LayoutTransformControl">
-        <Setter Property="RenderTransformOrigin" Value="50%,0%"/>
-        <Style.Animations>
-            <Animation Duration="0:0:0.75" Easing="QuadraticEaseOut" FillMode="Forward">
-                <KeyFrame Cue="0%">
-                    <Setter Property="TranslateTransform.X" Value="0"/>
-                    <Setter Property="ScaleTransform.ScaleY" Value="1"/>
-                </KeyFrame>
-                <KeyFrame Cue="70%">
-                    <Setter Property="TranslateTransform.X" Value="800"/>
-                    <Setter Property="ScaleTransform.ScaleY" Value="1"/>
-                </KeyFrame>
-                <KeyFrame Cue="100%">
-                    <Setter Property="ScaleTransform.ScaleY" Value="0"/>
-                    <Setter Property="TranslateTransform.X" Value="800"/>
-                </KeyFrame>
-            </Animation>
-        </Style.Animations>
-    </Style>
-
-    <Style Selector="Notification[IsClosing=true]">
-        <Style.Animations>
-            <Animation Duration="0:0:1.25" Easing="QuadraticEaseOut" FillMode="Forward">
-                <KeyFrame Cue="100%">
-                    <Setter Property="IsClosed" Value="True"/>
-                </KeyFrame>
-            </Animation>
-        </Style.Animations>
-    </Style>
-
-    <Style Selector="Notification /template/ Button#PART_CloseButton">
-        <Setter Property="Template">
-            <Setter.Value>
-                <ControlTemplate TargetType="Button">
-                    <TextBlock Text="X" FontSize="12" />
-                </ControlTemplate>
-            </Setter.Value>
-        </Setter>
-        <Setter Property="HorizontalAlignment" Value="Right"/>
-        <Setter Property="VerticalAlignment" Value="Top"/>
-    </Style>
-
-    <Style Selector="Notification">
-        <Setter Property="Background" Value="#444444"/>
-    </Style>
-    <Style Selector="Notification:information">
-        <Setter Property="Background" Value="Teal"/>
-    </Style>
-    <Style Selector="Notification:success">
-        <Setter Property="Background" Value="LimeGreen"/>
-    </Style>
-    <Style Selector="Notification:warning">
-        <Setter Property="Background" Value="Orange"/>
-    </Style>
-    <Style Selector="Notification:error">
-        <Setter Property="Background" Value="OrangeRed"/>
-    </Style>
-
     <Style Selector="WindowNotificationManager:topleft /template/ StackPanel#PART_Items">
         <Setter Property="VerticalAlignment" Value="Top"/>
         <Setter Property="HorizontalAlignment" Value="Left"/>
